pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastpki"
version = "0.1.0"
description = "API-based PKI management system"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
    { name = "FastPKI Team" },
]
dependencies = [
    "fastapi>=0.100.0",
    "sqlmodel>=0.0.8",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.22.0",
    "cryptography>=41.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "alembic>=1.11.0",
    "psycopg2-binary>=2.9.6",
    "asyncpg>=0.28.0",
    "aiosqlite>=0.19.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.27.0",
    "ruff>=0.1.6",
    "mypy>=1.5.0",
    "types-cryptography>=3.3.23.2",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py39"

# Linting rules
[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "TRY",  # tryceratops
    "RUF",  # ruff-specific rules
    "SIM",  # flake8-simplify
]
ignore = []

[tool.ruff.lint.isort]
known-third-party = ["fastapi", "pydantic", "sqlmodel"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]

# Formatting rules
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.9"
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
pretty = true
disallow_any_generics = true
disallow_untyped_defs = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
plugins = ["sqlalchemy.ext.mypy.plugin"]
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "app.api.*"
disallow_untyped_defs = false
disable_error_code = ["return-value", "arg-type"]

[[tool.mypy.overrides]]
module = "app.services.*"
disallow_untyped_defs = true
disable_error_code = ["return-value"]

[[tool.mypy.overrides]]
module = "app.db.*"
disable_error_code = ["call-overload"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "--cov=app --cov-report=term-missing --cov-report=xml"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"